Separate arch-dependent code to its own file
Code to create the seccomp filter is mostly arch-agnostic
(except for the amd64-specific header) but let's wait
with splitting it until an actual second architecture comes.
