+++ pdig.cc
@@ -24,6 +24,11 @@
 #include <atomic>
 #include <unordered_map>
 
+#include "pdig_api.h"
+#include "pdig_debug.h"
+#include "pdig_ptrace.h"
+#include "pdig_seccomp.h"
+
 
 enum class process_state {
 	spawning,
@@ -65,13 +70,6 @@ struct pdig_context {
 
 static std::atomic<bool> die(false);
 
-extern "C" {
-#include "pdig.h"
-#include "udig_capture.h"
-#include "ppm_events_public.h"
-#include "ppm.h"
-}
-
 void ignore_sig(int _sig)
 {
 }
@@ -92,22 +90,6 @@ static constexpr void* ptrace_options(bool use_seccomp)
 	return (void*)(use_seccomp ? SECCOMP_FLAGS : PTRACE_FLAGS);
 }
 
-
-pid_t get_pid(pid_t tid)
-{
-	struct user_regs_struct saved_regs;
-
-	EXPECT(ptrace(PTRACE_GETREGS, tid, &saved_regs, &saved_regs));
-
-	struct user_regs_struct regs = saved_regs;
-	regs.rax = __NR_getpid;
-	regs.rip = saved_regs.rip - 2;
-	EXPECT(ptrace(PTRACE_SETREGS, tid, &regs, &regs));
-	EXPECT(step(tid));
-	EXPECT(ptrace(PTRACE_GETREGS, tid, &regs, &regs));
-	EXPECT(ptrace(PTRACE_SETREGS, tid, &saved_regs, &saved_regs));
-	return regs.rax;
-}
 
 static bool wait_for_next_syscall(pid_t tid, pdig_process_context& pctx)
 {
@@ -136,42 +118,35 @@ static bool deliver_signal(pid_t tid, pdig_process_context& pctx, int sig)
 void handle_syscall(pid_t pid, pdig_process_context& pctx, bool enter)
 {
 	struct user_regs_struct regs;
-	uint64_t context[CTX_PID_TID + 1];
+	uint64_t context[CTX_SIZE];
 
 	if (ptrace(PTRACE_GETREGS, pid, &regs, &regs) < 0) {
 		WARN("PTRACE_GETREGS failed for pid %d", pid);
 		return;
 	}
-	context[CTX_REG_RDI] = regs.rdi;
-	context[CTX_REG_RSI] = regs.rsi;
-	context[CTX_REG_RDX] = regs.rdx;
-	context[CTX_REG_R10] = regs.r10;
-	context[CTX_REG_R8]  = regs.r8;
-	context[CTX_REG_R9]  = regs.r9;
-	context[CTX_REG_RAX_ENTER] = regs.orig_rax;
-	context[CTX_REG_RAX] = regs.rax;
-	context[CTX_PID_TID] = ((uint64_t)pctx.pid) << 32 | pid;
+	fill_context(context, pid, pctx.pid, regs);
 
-	DEBUG("pid=%d tid=%d syscall=%lld ret=%lld enter=%d rip=%016llx\n", pctx.pid, pid, regs.orig_rax, regs.rax, enter, regs.rip);
+	DEBUG("pid=%d tid=%d syscall=%ld ret=%ld enter=%d\n",
+		pctx.pid, pid, context[CTX_SYSCALL_ID], context[CTX_RETVAL], enter);
 
-	if(regs.orig_rax == __NR_rt_sigreturn)
+	if(context[CTX_SYSCALL_ID] == __NR_rt_sigreturn)
 	{
 		DEBUG("rt_sigreturn, faking exit event\n");
 		on_syscall(context, false);
 	}
-	else if(regs.orig_rax == __NR_execve && regs.rax == 0)
+	else if(context[CTX_SYSCALL_ID] == __NR_execve && context[CTX_RETVAL] == 0)
 	{
 		DEBUG("ignoring execve return, will capture PTRACE_EVENT_EXEC\n");
 	}
-	else if((int64_t)regs.orig_rax >= 0)
+	else if((int64_t)context[CTX_SYSCALL_ID] >= 0)
 	{
 		// rt_sigreturn yields two events, the other one with orig_rax == -1, ignore it
 		on_syscall(context, enter);
 
-		if(regs.orig_rax == __NR_clone && enter)
+		if(context[CTX_SYSCALL_ID] == __NR_clone && enter)
 		{
-			DEBUG("SYSCALL tid %d clone(%08llx)\n", pid, regs.rdi);
-			pctx.parent_clone_flags = regs.rdi;
+			DEBUG("SYSCALL tid %d clone(%08lx)\n", pid, context[CTX_ARG0]);
+			pctx.parent_clone_flags = context[CTX_ARG0];
 		}
 	}
 }
@@ -179,9 +154,9 @@ void handle_syscall(pid_t pid, pdig_process_context& pctx, bool enter)
 
 static bool handle_execve_exit(pid_t tid, pdig_process_context& pctx)
 {
-	uint64_t context[CTX_PID_TID + 1] = {0};
-	context[CTX_REG_RAX_ENTER] = __NR_execve;
-	context[CTX_REG_RAX] = 0;
+	uint64_t context[CTX_SIZE] = {0};
+	context[CTX_SYSCALL_ID] = __NR_execve;
+	context[CTX_RETVAL] = 0;
 	context[CTX_PID_TID] = ((uint64_t)tid) << 32 | tid; // pid == tid here
 	on_syscall(context, false);
 	return wait_for_next_syscall(tid, pctx);
@@ -217,14 +192,14 @@ static bool handle_ptrace_clone_event(pid_t tid, pdig_process_context& pctx, pdi
 static bool handle_spawning(pid_t tid, pdig_process_context& pctx)
 {
 	EXPECT(ptrace(PTRACE_SETOPTIONS, tid, 0, ptrace_options(pctx.use_seccomp)));
-	pctx.pid = get_pid(tid);
+	pctx.pid = inject_getpid(tid);
 
 	DEBUG("hello tid %d (pid %d)\n", tid, pctx.pid);
 	if(pctx.clone_flags) {
-		uint64_t context[CTX_PID_TID + 1] = {0};
-		context[CTX_REG_RAX_ENTER] = __NR_clone;
-		context[CTX_REG_RAX] = 0;
-		context[CTX_REG_RDI] = pctx.clone_flags;
+		uint64_t context[CTX_SIZE] = {0};
+		context[CTX_SYSCALL_ID] = __NR_clone;
+		context[CTX_RETVAL] = 0;
+		context[CTX_ARG0] = pctx.clone_flags;
 		context[CTX_PID_TID] = ((uint64_t)pctx.pid) << 32 | tid;
 		DEBUG("clone pid_tid = %016lx flags = %08lx\n", context[CTX_PID_TID], pctx.clone_flags);
 		on_syscall(context, false);
@@ -236,7 +211,7 @@ static bool handle_spawning(pid_t tid, pdig_process_context& pctx)
 static bool handle_attaching(pid_t tid, pdig_process_context& pctx)
 {
 	EXPECT(ptrace(PTRACE_SETOPTIONS, tid, 0, ptrace_options(pctx.use_seccomp)));
-	pctx.pid = get_pid(tid);
+	pctx.pid = inject_getpid(tid);
 	EXPECT(ptrace(PTRACE_SYSCALL, tid, NULL, NULL));
 
 	if(!pctx.use_seccomp) {
@@ -260,24 +235,11 @@ static bool handle_attaching_first_syscall_enter(pid_t tid, pdig_process_context
 
 static bool handle_attaching_first_syscall_exit(pid_t tid, pdig_process_context& pctx, pdig_context& main_ctx)
 {
-	struct sock_fprog* prog = main_ctx.seccomp_filter;
 	struct user_regs_struct saved_regs;
 
 	EXPECT(ptrace(PTRACE_GETREGS, tid, &saved_regs, &saved_regs));
 
-	struct user_regs_struct regs = saved_regs;
-	regs.rax = __NR_prctl;
-	regs.rdi = PR_SET_NO_NEW_PRIVS;
-	regs.rsi = 1;
-	regs.rdx = 0;
-	regs.r10 = 0;
-	regs.r8 = 0;
-	regs.rip = saved_regs.rip - 2;
-	EXPECT(ptrace(PTRACE_SETREGS, tid, &regs, &regs));
-	EXPECT(step(tid));
-	EXPECT(ptrace(PTRACE_GETREGS, tid, &regs, &regs));
-	if(regs.rax != 0) {
-		WARN("Failed to enable PR_SET_NO_NEW_PRIVS, disabling seccomp%s", "");
+	if (inject_prctl_set_no_new_privs(tid, saved_regs) != 0) {
 		pctx.use_seccomp = false;
 		EXPECT(ptrace(PTRACE_SETREGS, tid, &saved_regs, &saved_regs));
 		EXPECT(ptrace(PTRACE_SYSCALL, tid, NULL, NULL));
@@ -285,35 +247,7 @@ static bool handle_attaching_first_syscall_exit(pid_t tid, pdig_process_context&
 		return true;
 	}
 
-	// set stack pointer before we copy our data over
-	EXPECT(ptrace(PTRACE_SETREGS, tid, &regs, &regs));
-
-	// stack layout
-	//
-	// [prog][sock_filter[]][red zone][orig stack]
-	// ^     ^                        ^
-	// |     |                        saved_regs.rsp
-	// |     sock_fprog->filter
-	// rdx, rsp
-
-	unsigned long payload_len = sizeof(*prog) + prog->len * sizeof(prog->filter[0]);
-	unsigned long payload_addr = (saved_regs.rsp - payload_len - 128) & ~0x0f;
-	regs = saved_regs;
-	regs.rax = __NR_prctl;
-	regs.rdi = PR_SET_SECCOMP;
-	regs.rsi = SECCOMP_MODE_FILTER;
-	regs.rdx = payload_addr;
-	prog->filter = (struct sock_filter*)(payload_addr + sizeof(*prog));
-	regs.r10 = 0;
-	regs.r8 = 0;
-	regs.rip = saved_regs.rip - 2;
-	regs.rsp = payload_addr;
-
-	copy_to_user(tid, prog, (void*)payload_addr, (payload_len + 0x0f) & ~0x0f);
-
-	EXPECT(step(tid));
-	EXPECT(ptrace(PTRACE_GETREGS, tid, &regs, &regs));
-	if(regs.rax != 0) {
+	if (inject_seccomp_filter(tid, saved_regs, main_ctx.seccomp_filter) != 0) {
 		abort();
 	}
 
@@ -430,118 +364,6 @@ static bool handle_waitpid(pid_t tid, int status, pdig_context& main_ctx)
 }
 
 
-
-#define X32_SYSCALL_BIT 0x40000000
-#define ARRAY_SIZE(a) (sizeof(a) / sizeof((a)[0]))
-
-static struct sock_fprog* build_filter(bool capture_all)
-{
-	size_t num_syscalls = 0;
-
-	uint32_t instrumented_syscalls[SYSCALL_TABLE_SIZE] = {0};
-
-	for(size_t i=0; i<SYSCALL_TABLE_SIZE; ++i) {
-		uint32_t flags = g_syscall_table[i].flags;
-		if (!(flags & UF_USED)) {
-			continue;
-		}
-
-		bool instrument = capture_all;
-
-		if(!instrument) {
-			enum ppm_event_type enter_event = g_syscall_table[i].enter_event_type;
-			enum ppm_event_type exit_event = g_syscall_table[i].exit_event_type;
-
-			uint32_t enter_event_flags = g_event_info[enter_event].flags;
-			uint32_t exit_event_flags = g_event_info[exit_event].flags;
-
-			instrument = (enter_event_flags & EF_MODIFIES_STATE) || (exit_event_flags & EF_MODIFIES_STATE);
-		}
-
-		if(instrument) {
-			instrumented_syscalls[num_syscalls++] = i;
-		}
-	}
-
-	struct sock_filter filter_header[] = {
-		// if arch != AUDIT_ARCH_X86_64 { return ERRNO | ENOSYS }
-		BPF_STMT(BPF_LD | BPF_W | BPF_ABS, (offsetof(struct seccomp_data, arch))),
-		BPF_JUMP(BPF_JMP | BPF_JEQ | BPF_K, AUDIT_ARCH_X86_64, 1, 0),
-		BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_ERRNO | (ENOSYS & SECCOMP_RET_DATA)),
-
-		// if nr >= X32_SYSCALL_BIT { return ERRNO | ENOSYS }
-		BPF_STMT(BPF_LD | BPF_W | BPF_ABS, (offsetof(struct seccomp_data, nr))),
-		BPF_JUMP(BPF_JMP | BPF_JGE | BPF_K, X32_SYSCALL_BIT, 0, 1),
-		BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_ERRNO | (ENOSYS & SECCOMP_RET_DATA)),
-	};
-
-	struct sock_filter filter_footer[] = {
-		BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_ALLOW),
-	};
-
-	const size_t CHUNK_SIZE = 255; // must be < 256
-
-	size_t num_chunks = (num_syscalls + CHUNK_SIZE - 1) / CHUNK_SIZE;
-	size_t last_chunk_len = num_syscalls % CHUNK_SIZE;
-	size_t trace_insn = ARRAY_SIZE(filter_header) - 1;
-
-	size_t filter_size = ARRAY_SIZE(filter_header) + num_syscalls + num_chunks + ARRAY_SIZE(filter_footer);
-	DEBUG("filter has %zu insns\n", filter_size);
-	size_t payload_size = sizeof(struct sock_fprog) + filter_size * sizeof(struct sock_filter);
-	char* buf = (char*)calloc((payload_size + 0x0f) & ~0x0f, 1);
-	if(!buf) {
-		return NULL;
-	}
-
-	struct sock_filter* filter = (struct sock_filter*)(buf + sizeof(struct sock_fprog));
-	struct sock_fprog* prog = (struct sock_fprog*)buf;
-
-	memcpy(&filter[0], filter_header, sizeof(filter_header));
-	int insn = ARRAY_SIZE(filter_header);
-
-	for(size_t i = 0; i < num_syscalls; ++i) {
-		size_t chunk = i / CHUNK_SIZE;
-		size_t chunk_off = i % CHUNK_SIZE;
-		size_t current_chunk_size;
-		if (chunk == num_chunks - 1) {
-			current_chunk_size = last_chunk_len;
-		} else {
-			current_chunk_size = CHUNK_SIZE;
-		}
-		int at_chunk_end = (chunk_off == current_chunk_size - 1);
-
-		if (chunk_off == 0) {
-			trace_insn += current_chunk_size + 1;
-			if (chunk != 0) {
-				DEBUG("[insn#%d] RET_TRACE\n", insn);
-				filter[insn++] = BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_TRACE);
-			}
-		}
-
-		int jt = trace_insn - insn - 1;
-		int jf = at_chunk_end;
-
-		if(jt != (uint8_t)jt) {
-			abort();
-		}
-
-		if(jf != (uint8_t)jf) {
-			abort();
-		}
-
-		DEBUG("[insn#%d] if nr == %d { goto +%d (%d) } else { goto +%d (%d) }\n",
-			insn, instrumented_syscalls[i], jt, insn + jt + 1, jf, insn + jf + 1);
-		filter[insn++] = BPF_JUMP(BPF_JMP | BPF_JEQ | BPF_K, instrumented_syscalls[i], (uint8_t)jt, (uint8_t)jf);
-	}
-	DEBUG("[insn#%d] RET_TRACE\n", insn);
-	filter[insn++] = BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_TRACE);
-
-	memcpy(&filter[insn], filter_footer, sizeof(filter_footer));
-
-	prog->len = (unsigned short)filter_size;
-	prog->filter = filter;
-	return prog;
-}
 
 static int spawn(int argc, char** argv, pdig_context& main_ctx)
 {
