+++ pdig_compat.h
@@ -1,0 +1,4 @@
+#ifndef PDIG_COMPAT_H
+#define PDIG_COMPAT_H
+
+#endif
